Fix Float/Decimal precision drift in consumption columns
raw_value/est_kwh were declared as Float, silently truncating the
Decimal values the domain model (data.model.Consumption) already
uses throughout, and diverging from the old init.sql's DECIMAL(8,5).
Changed to Numeric(8,5) so persisted values round-trip as Decimal
with no precision loss. Removes a Decimal(str(...)) workaround in
test_consumption_seam.py that existed specifically because of this
bug, and adds a regression test proving round-trip precision.

Addresses an advisory finding from the schema-sync code review;
fixed now rather than deferred since nothing has been deployed yet.

+++ app/data/mysql/sql_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Float, Integer, Numeric, String
+from sqlalchemy import Column, DateTime, Integer, Numeric, String
 from sqlalchemy.ext.declarative import declarative_base
 
 SQLBase = declarative_base()
@@ -12,9 +12,9 @@ class consumption(SQLBase):
     energy = Column(String(1))
     period_from = Column(DateTime, nullable=False)
     period_to = Column(DateTime, nullable=False)
-    raw_value = Column(Float, nullable=False)
+    raw_value = Column(Numeric(8, 5), nullable=False)
     unit = Column(String(5))
-    est_kwh = Column(Float, nullable=False)
+    est_kwh = Column(Numeric(8, 5), nullable=False)
 
 
 class agreement(SQLBase):
